refactor(food): definitions/fruit.tsをfoodディレクトリ配下に移動
