+++ src/definitions/food/fruit.ts
