+++ src/modules/food/fruit.test.ts
@@ -1,5 +1,5 @@
 import { fruit } from './fruit'
-import { fruit as fruitDefinitions } from '../../definitions/fruit'
+import { fruit as fruitDefinitions } from '../../definitions/food/fruit'
 
 describe('fruit', () => {
   it('果物の配列から値を返すこと', () => {
